Remove mempool clear command
