+++ monad-ipc/examples/ipc_receiver.rs
@@ -147,7 +147,7 @@ async fn main() -> Result<(), Box<dyn Error>> {
                         debug!("received tx");
                     }
                 }
-                MempoolEvent::ForwardedTxns { .. } | MempoolEvent::Clear => {}
+                MempoolEvent::ForwardedTxns { .. } => {}
             },
 
             _ => Err("Wrong MonadEvent variant")?,
